requirements.txt
Django>=2
psycopg2-binary
djangorestframework
celery
redis
django-admin-json-editor